requirements.txt
# Core dependencies
mss>=9.0.1              # Cross-platform screenshot capture
pynput>=1.7.6            # Cross-platform input monitoring
pillow>=10.0.0           # Image processing
numpy>=1.24.0            # Array operations
opencv-python>=4.8.0     # Computer vision and image processing

# OCR dependencies
pytesseract>=0.3.10      # OCR text extraction
easyocr>=1.7.0           # Backup OCR engine

# Web server for editing interface
flask>=2.3.0             # Local web server
flask-cors>=4.0.0        # CORS support for localhost

# Document export
python-docx>=0.8.11      # Word document export
reportlab>=4.0.0         # PDF generation

# Desktop GUI dependencies
pystray>=0.19.4          # System tray integration
keyboard>=0.13.5         # Global hotkeys (optional)

# Utilities
uuid>=1.30               # Session ID generation
watchdog>=3.0.0          # File system monitoring
typing-extensions>=4.7.0 # Type hints support

# Note: tkinter is included with Python standard library